Use more specific fuzzy matching limits for 2d.composite.grid tests
These tests are all implemented as a single parameterized test. The
the fuzzy matching was defined in the base test definition, which means
that all variants of the parameterized tests we using the same fuzzy
config. This unfortunately required all the variants to share the same
worse-case fuzzy matching limits.

With a little Jinja templating, we can be more granular with how we
specify that config. In this CL, the fuzzy config is specified using a
map that is keyed based on the variant name, allowing each variant to
have different config.

This will be useful in a follow-up CL where adding blur to the shadow
tests will further increase the difference between the fuzzy matching
limits of different variants.

Bug: 364549423
Change-Id: I20ada8d1b1a97b07b1a6eb5c4b3b98a8c8fcf9ac